Update support for ACPI 5.0 MPST table.
Fixes some problems introduced by late changes to the table as it
was added to the ACPI 5.0 specification.
Both the table compiler and the disassembler and the main header
support for the table.

+++ source/compiler/dttable.c
@@ -1256,8 +1256,8 @@ DtCompileMpst (
     ACPI_MPST_POWER_NODE    *MpstPowerNode;
     ACPI_MPST_DATA_HDR      *MpstDataHeader;
     UINT16                  SubtableCount;
-    UINT8                   PowerStateCount;
-    UINT8                   ComponentCount;
+    UINT32                  PowerStateCount;
+    UINT32                  ComponentCount;
 
 
     /* Main table */
